Fix DSPFFD out of GENCMDXML and into utils
